[major] replace node-sass with stylus as css preprocessor
This switch is mainly due to unstableness of node-sass cross different platforms. Since stylus is syntactically very similar to sass, and imo more capable than sass in many way, I see no reason against this switch. Conversion between the two lang is pretty easy, since stylus can also be used as a superset of css, plus stylus is grammatically more close to js, it actually save a lot of cognitive cost in adopting another lang.

However, in order to facilitate adoption of those who aren't familiar with this lang, the (awesomely) concise syntax of stylus is intentionally not used (and will continue to be discouraged) in this project. Things are kept mostly aligned with plain old css.